ahci: fix command timeout 1h11m36s after VMM starts
Previously the ahci_command_completion function had a 32bit
argument for current time.  The time of the limit of 32bit,
4294967295, is about 4295 seconds, 1h11m35s.  After that time,
the start time which was stored in a 64bit variable was always
larger than the 32bit time.  This changeset fixes the bug by
changing the argument to 64bit.

+++ drivers/ata/ahci.c
@@ -1283,7 +1283,7 @@ found:
 
 static bool
 ahci_command_completion (struct ahci_data *ad, struct ahci_command_list *p,
-			 struct ahci_command_data *data, u32 time)
+			 struct ahci_command_data *data, u64 time)
 {
 	struct recvfis *fis;
 	struct storage_hc_dev_atacmd *cmd;
